% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip-history.R
\name{ip_history}
\alias{ip_history}
\title{Retrieve IP ASN advertisement history}
\usage{
ip_history(ip, source = "caida", address_family = c("v4", "v6"),
  date = NULL, first = NULL, last = NULL, precision_delta = NULL,
  ip_asn_history_endpoint = "https://bgpranking-ng.circl.lu/ipasn_history/")
}
\arguments{
\item{ip}{(required) IP to lookup}

\item{source}{(optional) Source to query (defaults to 'caida') - currently, only caida is supported}

\item{address_family}{(optional) \code{v4} or \code{v6} (defaults to v4)}

\item{date}{(optional) Exact date to lookup (defaults to most recent available). Can
be a date-y/time-y-ish R object or a "YYYY-mm-dd" string}

\item{first, last}{(optional) First/last date in the interval. Same value format restrictions as \code{date} param}

\item{precision_delta}{(optional) Max delta allowed between the date queried and the one we have in the database.
Use \code{\link[=timedelta]{timedelta()}} to construct the value for this parameter.}

\item{ip_asn_history_endpoint}{supply or your own or use the default circl.lu
server provided.}
}
\description{
Retrieve IP ASN advertisement history
}
\note{
If you want to resolve more than 1000 IPs, use \code{\link[=mass_cache]{mass_cache()}} and \code{\link[=mass_query]{mass_query()}}
}
\examples{
ip_history("146.185.222.49")
ip_history("146.185.222.49", first="2018-11-01", last="2018-11-25")
}
