% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipasnhistory-package.R
\docType{package}
\name{ipasnhistory}
\alias{ipasnhistory}
\alias{ipasnhistory-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/ipasnhistory}
\item BugReports: \url{https://gitlab.com/hrbrmstr/ipasnhistory/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
