% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-query.R
\name{mass_query}
\alias{mass_query}
\title{Mass query IP ASN history info}
\usage{
mass_query(ips, source = NULL, address_family = NULL, date = NULL,
  first = NULL, last = NULL, precision_delta = NULL,
  ip_asn_history_endpoint = "https://bgpranking-ng.circl.lu/ipasn_history/")
}
\arguments{
\item{ips}{(required) character vector of IPs to lookup}

\item{source}{(optional) Source to query (defaults to 'caida') - currently, supports CAIDA & RIPE NCC sources}

\item{address_family}{(optional) \code{v4} or \code{v6} (defaults to v4)}

\item{date}{(optional) Exact date to lookup (defaults to most recent available). Can
be a date-y/time-y-ish R object or a "YYYY-mm-dd" string}

\item{first, last}{(optional) First/last date in the interval. Same value format restrictions as \code{date} param}

\item{precision_delta}{(optional) Max delta allowed between the date queried and the one we have in the database.
Use \code{\link[=timedelta]{timedelta()}} to construct the value for this parameter.}

\item{ip_asn_history_endpoint}{supply or your own or use the default circl.lu
server provided.}
}
\description{
Use this path when you have lots of query to run and (>1000)
in order to resolve all of them at once and wait for the results.
See Details for how to specify parameter values
}
\details{
This is a bulk cache/lookup function. If \code{source}, \code{address_family}, \code{date},
\code{first}, \code{last}, \code{precision_delta} are not either \code{NULL} or length 1 (
for \code{precision_delta} that means a single list retrieved from \code{\link[=timedelta]{timedelta()}})
then they \emph{must be the same length as \code{ips}}, since the API expects
a full parameter list for each IP being queried for and there's no way
for this function to intuit what you need outside of these specified rules.
}
\note{
If using your own server do not add the \code{/mass_query} at the end of the
\code{ip_asn_history_endpoint} URL parameter.
}
