% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-delta.R
\name{timedelta}
\alias{timedelta}
\title{Construct a time delta value for \code{\link[=ip_history]{ip_history()}}}
\usage{
timedelta(days = 0, seconds = 0, microseconds = 0,
  milliseconds = 0, minutes = 0, hours = 0, weeks = 0)
}
\arguments{
\item{days, seconds, microseconds, milliseconds, minutes, hours, weeks}{time delta
components. All default to \code{0} so just change what you need which
will keep the function call fairly short in most cases.}
}
\description{
Construct a time delta value for \code{\link[=ip_history]{ip_history()}}
}
