% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipasnhistory-package.R
\docType{package}
\name{ipasnhistory}
\alias{ipasnhistory}
\alias{ipasnhistory-package}
\title{Retrieve Announcement History for IP Addresses}
\description{
IP ASN History is an open source project (\url{https://github.com/D4-project/IPASN-History})
to figure out the ASN announcing an IP and the closest prefix announcing it
at a specific date. Methods are provided to query CIRCL-LU's API endpoint
and/or a local API endpoint.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/ipasnhistory}
\item BugReports: \url{https://gitlab.com/hrbrmstr/ipasnhistory/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
